requirements.txt
asgiref==3.3.1
dj-static==0.0.6
Django==3.1.4
django-crispy-forms==1.10.0
django-stdimage==5.1.1
mysqlclient==2.0.1
phonenumbers==8.12.15
Pillow==8.0.1
progressbar2==3.53.1
python-http-client==3.3.1
python-utils==2.4.0
pytz==2020.4
sendgrid==3.6.5
sendgrid-django==4.2.0
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
validate-docbr==1.8.2
